fixing sql queries

+++ delete.php
@@ -20,11 +20,17 @@ if (isset($_POST['password']) && PASSWORD != sha1($_POST['password'])) {
 
     $dq = $pdo->query('SELECT COUNT(*) FROM news WHERE id='.$pdo->quote($id),PDO::FETCH_UNIQUE);
     $re = $dq->fetch();
-
+    $re = array_shift($re);
     if ($re == 1) {
         $pdo->prepare('DELETE FROM news WHERE id=?');
-        $dq->execute(array(intval($_POST['id'])));
-        unlink('cache/n-'.$id);
+        $q = $dq->execute(array(intval($_POST['id'])));
+        if (!$q) {
+            $text = "The news is not deleted : " . $pdo->errorInfo();
+        } else {
+            $text = "The news $id is successfully deleted.<br/><a href=\"index\">Back</a>";
+            unlink('cache/n-'.$id);
+            unlink('cache/n-'.$id.'.minimal');
+        }
     }
 }
 ?>
